cli-client/pyproject.toml
[tool.poetry]
name = "player-ranking"
version = "0.1.0"
description = "CLI interface to compute a ranking between all players in a Clash Royale clan"
authors = ["Vincent Raudszus"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
requests = "^2.32.3"
pandas = "^2.2.2"
matplotlib = "^3.9.2"
matplotlib-label-lines = "^0.7.0"
google-api-python-client = "^2.143.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
python-dotenv = "^1.0.1"
jsonschema = "^4.23.0"

[tool.poetry.scripts]
player-ranking = "run_player_ranking:run"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
pre-commit = "^3.8.0"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
